Replaced PlTracks with Songs
Instead of having a separate PlTrack class for elements of a playlist,
I made playlists a list of Songs, with an additional number element
denoting their position in the playlist.

For the record, I want number to be replaced with playlist_order or
something else so it doesn't break everything. Not sure if that will break
compatibility though.